adding `switch-user` gulp task
this task is specific to a workflow where the script is not an add-on but is used by multiple accounts (therefore a separate auth is needed for each account). The credentials must be stored in a dir named `credentials`, file name must match the options.user variable supplied when the task is called.